Adjust some spacing and code organization
